Turned off all warnings for the Pods frameworks. Either commented out things in our project that need to be worked on that are causing warnings or fixed them.

+++ Ignite_iOS_Engine/Classes/Model/Properties/ShortCodes/IXAppShortCode.m
@@ -32,7 +32,7 @@
         {
             upperBound = [[stringArray lastObject] intValue];
         }
-        returnValue = [NSString stringWithFormat:@"%i",arc4random_uniform(upperBound)];
+        returnValue = [NSString stringWithFormat:@"%i",arc4random_uniform((u_int32_t)upperBound)];
     }
     
     return returnValue;
